From: jan.dubois@ibm.net (Jan Dubois) Subject: [PATCH] Unicode-String-2.02: compile with C++ (PERL_OBJECT) To: Gisle Aas <gisle@aas.no> Date: Wed, 05 May 1999 22:35:21 +0200
Hi Gisle,

I just tried to compile Unicode-String under ActivePerl 515.
It compiles the module in C++ mode because of the PERL_OBJECT
option. The String.xs code has two problems: It uses 'new' as
a variable name, which is a reserved word in C++. And some
signed/unsigned casts are missing.

With the attached patch applied, all tests pass.